pyproject.toml
[project]
name = "asistencia-pyro"
version = "0.1.0"
description = ""
authors = [
    {name = "“Forek4st”",email = "“janeirojm91@gmail.com”"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "tortoise-orm (>=0.25.1,<0.26.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.7,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
